Split review UI into focused components
