+++ src/ui/review-status-line.tsx
@@ -1,0 +1,9 @@
+export function StatusLine({
+  message,
+  tone = "muted",
+}: {
+  message: string;
+  tone?: "muted" | "error";
+}) {
+  return <div className={`status ${tone}`}>{message}</div>;
+}
